chore: create separate folders for reusable components (sections)
