+++ client/src/pages/karriar/index.js
@@ -3,7 +3,8 @@ import { groq } from 'next-sanity'
 import { NextSeo } from 'next-seo'
 
 import PageWrapper from '../../components/PageWrapper'
-import { Hero, Testimonial } from '../../sections/common'
+import { Hero } from '../../sections/Hero'
+import { Testimonial } from '../../sections/Testimonial'
 import { Feature, Roles } from '../../sections/Karriar'
 import { usePreviewSubscription } from '../../lib/sanity'
 import { getClient } from '../../lib/sanity.server'
