+++ client/src/sections/Pricing/index.js
@@ -1,0 +1,1 @@
+export { Pricing } from './Pricing'
