+++ client/src/sections/SectionWithButtonAndTextGrid/index.js
@@ -1,0 +1,1 @@
+export { SectionWithButtonAndTextGrid } from './SectionWithButtonAndTextGrid'
