+++ client/src/sections/SectionWithImage/SectionWithImage.js
@@ -12,7 +12,7 @@ const SectionStyled = styled(Section)`
   background-color: ${({ bg }) => (bg == 'dark' ? 'rgb(25,25,27)' : 'white')};
   overflow: hidden;
 `
-const SectionWithImage = ({ content, bg = 'light' }) => {
+export const SectionWithImage = ({ content, bg = 'light' }) => {
   return (
     <SectionStyled bg={bg}>
       <Container>
@@ -60,5 +60,3 @@ const SectionWithImage = ({ content, bg = 'light' }) => {
     </SectionStyled>
   )
 }
-
-export default SectionWithImage
