+++ client/src/sections/SectionWithImage/index.js
@@ -1,0 +1,1 @@
+export { SectionWithImage } from './SectionWithImage'
