+++ client/src/sections/SectionWithImageAndCta/index.js
@@ -1,0 +1,1 @@
+export { SectionWithImageAndCta } from './SectionWithImageAndCta'
