+++ client/src/sections/Testimonial/index.js
@@ -1,0 +1,1 @@
+export { Testimonial } from './Testimonial'
