+++ client/src/sections/TextWithTags/index.js
@@ -1,0 +1,1 @@
+export { TextWithTags } from './TextWithTags'
