--- client/src/sections/common/index.js
@@ -1,14 +1,0 @@
-import SectionWithButtonAndTextGrid from './SectionWithButtonAndTextGrid'
-import SectionWithImage from './SectionWithImage'
-import TextWithTags from './TextWithTags'
-import Testimonial from './Testimonial'
-import Hero from './Hero'
-import { Pricing } from './Pricing'
-export {
-  SectionWithButtonAndTextGrid,
-  SectionWithImage,
-  TextWithTags,
-  Testimonial,
-  Hero,
-  Pricing,
-}
